pyproject.toml
[project]
name = "myrio"
version = "0.0.1"
description = "A pipeline for taxonomic identification of plants using sequencing data."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiofiles>=24.1.0",
    "biopython>=1.85",
    "clypi>=1.2.18",
    "isonclust>=0.0.6.1",
    "numpy>=2.2.5",
    "polars>=1.27.1",
    "safe-result>=4.0.3",
]

[project.scripts]
myrio = "cli:main"

[tool.uv]
dev-dependencies = [
    "pytest-benchmark>=5.1.0",
    "pytest>=8.3.5",
    "ruff>=0.11.6",
    "pyclean>=3.1.0",
    "coverage>=7.8.0",
    "codespell>=2.4.1",
    "lefthook>=1.11.11",
    "marimo>=0.13.2",
    "seaborn>=0.13.2",
    "matplotlib>=3.10.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build]
sources = ["src"]

[tool.hatch.build.targets.wheel]
include = ["*.py"]

[tool.pytest.ini_options]
addopts = "--benchmark-skip"
console_output_style = "count"
python_functions = "test_* *_test bench_* *_bench"
python_files = "test_*.py *_test.py bench_*.py *_bench.py"

[tool.ruff]
line-length = 120

[tool.ruff.lint]
extend-select = ["I"]
ignore = ["E702", "F841"] # Allows for multiple statements on a single line

[tool.pyright]
include = ["src"]
exclude = ["notebooks", "ignore"]
